software_architecture_mvp_backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_migrate
bcrypt
pyjwt
python-dotenv
requests